added the option to disable the skip empty blocks feature
