+++ block/manager_test.go
@@ -5,7 +5,6 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"errors"
-	"fmt"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -16,7 +15,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dymensionxyz/dymint/log/test"
-	"github.com/dymensionxyz/dymint/mempool"
 	mempoolv1 "github.com/dymensionxyz/dymint/mempool/v1"
 	"github.com/dymensionxyz/dymint/node/events"
 	"github.com/dymensionxyz/dymint/p2p"
@@ -25,7 +23,6 @@ import (
 	"github.com/dymensionxyz/dymint/types"
 	"github.com/libp2p/go-libp2p-core/crypto"
 	abci "github.com/tendermint/tendermint/abci/types"
-	cfg "github.com/tendermint/tendermint/config"
 	tmcfg "github.com/tendermint/tendermint/config"
 	"github.com/tendermint/tendermint/libs/log"
 	"github.com/tendermint/tendermint/libs/pubsub"
@@ -564,72 +561,6 @@ func TestCreateNextDABatchWithBytesLimit(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestCreateEmptyBlocksNew(t *testing.T) {
-	assert := assert.New(t)
-	require := require.New(t)
-	app := testutil.GetAppMock()
-	// Create proxy app
-	clientCreator := proxy.NewLocalClientCreator(app)
-	proxyApp := proxy.NewAppConns(clientCreator)
-	err := proxyApp.Start()
-	require.NoError(err)
-	// Init manager
-	managerConfig := getManagerConfig()
-	managerConfig.BlockTime = 200 * time.Millisecond
-	managerConfig.EmptyBlocksMaxTime = 1 * time.Second
-	manager, err := getManager(managerConfig, nil, nil, 1, 1, 0, proxyApp, nil)
-	require.NoError(err)
-
-	abciClient, err := clientCreator.NewABCIClient()
-	require.NoError(err)
-	require.NotNil(abciClient)
-	require.NoError(abciClient.Start())
-	defer abciClient.Stop()
-
-	mempoolCfg := cfg.DefaultMempoolConfig()
-	mempoolCfg.KeepInvalidTxsInCache = false
-	mempoolCfg.Recheck = false
-
-	mpool := mempoolv1.NewTxMempool(log.TestingLogger(), cfg.DefaultMempoolConfig(), proxy.NewAppConnMempool(abciClient), 0)
-
-	//Check initial height
-	expectedHeight := uint64(0)
-	assert.Equal(expectedHeight, manager.store.Height())
-
-	mCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	go manager.ProduceBlockLoop(mCtx)
-
-	<-time.Tick(1 * time.Second)
-	err = mpool.CheckTx([]byte{1, 2, 3, 4}, nil, mempool.TxInfo{})
-	require.NoError(err)
-
-	<-mCtx.Done()
-	foundTx := false
-	assert.LessOrEqual(manager.store.Height(), uint64(10))
-	for i := uint64(2); i < manager.store.Height(); i++ {
-		prevBlock, err := manager.store.LoadBlock(i - 1)
-		assert.NoError(err)
-
-		block, err := manager.store.LoadBlock(i)
-		assert.NoError(err)
-		assert.NotZero(block.Header.Time)
-
-		diff := time.Unix(0, int64(block.Header.Time)).Sub(time.Unix(0, int64(prevBlock.Header.Time)))
-		txsCount := len(block.Data.Txs)
-		if txsCount == 0 {
-			assert.Greater(diff, manager.conf.EmptyBlocksMaxTime)
-			assert.Less(diff, manager.conf.EmptyBlocksMaxTime+1*time.Second)
-		} else {
-			foundTx = true
-			assert.Less(diff, manager.conf.BlockTime+100*time.Millisecond)
-		}
-
-		fmt.Println("time diff:", diff, "tx len", 0)
-	}
-	assert.True(foundTx)
 }
 
 /* -------------------------------------------------------------------------- */
